recreate list item
