+++ docs/src/pages/components/List.vue
@@ -957,15 +957,7 @@
   }
 
   .custom-list {
-    .md-list-action {
-      position: absolute;
-      top: 8px;
-      right: 16px;
-      pointer-events: auto;
-      z-index: 2;
-    }
-
-    .md-icon {
+    .md-icon:not(.md-primary) {
       color: rgba(#000, .26);
     }
   }
