+++ src/components/mdList/mdList.scss
@@ -114,10 +114,6 @@
     .md-avatar {
       margin: 0;
     }
-
-    .md-list-item-container {
-      align-items: flex-start;
-    }
   }
 
   .md-subheader {
@@ -134,6 +130,12 @@
 .md-list-item {
   height: auto;
   position: relative;
+  z-index: 2;
+
+  &.md-disabled {
+    cursor: default;
+    pointer-events: none;
+  }
 
   &.md-inset {
     .md-list-item-container {
@@ -141,15 +143,44 @@
     }
   }
 
-  .md-list-item-holder {
+  .md-button-ghost {
+    width: 100%;
+    margin: 0;
+    position: absolute;
+    top: 0;
+    right: 0;
+    bottom: 0;
+    left: 0;
+    z-index: 1;
+    border-radius: 0;
+  }
+
+  .md-button:not(.md-button-ghost):not(.md-list-item-container) {
+    position: relative;
+    z-index: 2;
+
+    .md-icon {
+      position: relative;
+    }
+  }
+
+  .md-list-item-container {
+    min-height: 48px;
+    margin: 0;
+    padding: 0 16px;
     display: flex;
     flex-flow: row nowrap;
     align-items: center;
     justify-content: space-between;
     flex: 1;
+    position: relative;
+    font-size: 16px;
+    font-weight: 400;
+    text-align: left;
+    text-transform: none;
 
-    > .md-ink-ripple {
-      border-radius: 0;
+    &:hover {
+      text-decoration: none;
     }
 
     > .md-icon:first-child {
@@ -161,25 +192,12 @@
     }
 
     .md-list-action {
-      margin: 0 -2px 0 0;
+      margin: 0 -10px 0 0;
 
       &:nth-child(3) {
-        margin: 0 -2px 0 16px;
+        margin: 0 -10px 0 16px;
       }
     }
-  }
-
-  .md-list-item-container {
-    width: 100%;
-    min-height: 48px;
-    margin: 0;
-    padding: 0 16px;
-    position: relative;
-    border-radius: 0;
-    font-size: 16px;
-    font-weight: 400;
-    text-align: left;
-    text-transform: none;
   }
 
   .md-divider {
@@ -216,6 +234,7 @@
   min-height: 48px;
   flex-flow: column wrap;
   overflow: hidden;
+  transform: translate3D(0, 0, 0);
 
   &:before,
   &:after {
@@ -265,11 +284,6 @@
     }
   }
 
-  > .md-list-item-container > .md-list-item-holder {
-    position: relative;
-    z-index: 2;
-  }
-
   .md-expansion-indicator,
   .md-list-item-container,
   .md-icon {
@@ -299,8 +313,13 @@
   flex: 1;
   overflow: hidden;
   line-height: 1.25em;
-  text-overflow: ellipsis;
   white-space: normal;
+
+  > * {
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+  }
 
   > :nth-child(1) {
     font-size: 16px;
