+++ src/components/mdList/mdListItemButton.vue
@@ -1,0 +1,25 @@
+<template>
+  <li class="md-list-item" :class="classes">
+    <div class="md-list-item-container md-button">
+      <slot></slot>
+    </div>
+
+    <md-button class="md-button-ghost" :disabled="disabled"></md-button>
+  </li>
+</template>
+
+<script>
+  export default {
+    name: 'md-list-item',
+    props: {
+      disabled: Boolean
+    },
+    computed: {
+      classes() {
+        return {
+          'md-disabled': this.disabled
+        };
+      }
+    }
+  };
+</script>
