+++ src/components/mdList/mdListItemDefault.vue
@@ -1,0 +1,13 @@
+<template>
+  <li class="md-list-item">
+    <div class="md-list-item-container">
+      <slot></slot>
+    </div>
+  </li>
+</template>
+
+<script>
+  export default {
+    name: 'md-list-item'
+  };
+</script>
