+++ src/core/stylesheets/base.scss
@@ -27,8 +27,13 @@ body {
   @extend .md-body-1;
 }
 
-[tabindex='-1']:focus {
-  outline: none;
+
+
+/*  Fluid Media
+   ========================================================================== */
+
+ul:not(.md-list) > li + li {
+  margin-top: 8px;
 }
